Add support for backend favorites

+++ js/views/roomlistview.js
@@ -43,6 +43,23 @@
 						'</div>'+
 						'<div class="app-navigation-entry-menu">'+
 							'<ul class="app-navigation-entry-menu-list">'+
+								'{{#if canFavorite}}'+
+								'{{#if isFavorite}}'+
+								'<li>'+
+									'<button class="unfavorite-room-button">'+
+										'<span class="icon-star-dark"></span>'+
+										'<span>'+t('spreed', 'Unpin conversation')+'</span>'+
+									'</button>'+
+								'</li>'+
+								'{{else}}'+
+								'<li>'+
+									'<button class="favorite-room-button">'+
+										'<span class="icon-starred"></span>'+
+										'<span>'+t('spreed', 'Pin conversation')+'</span>'+
+									'</button>'+
+								'</li>'+
+								'{{/if}}'+
+								'{{/if}}'+
 								'{{#if isRemovable}}'+
 								'<li>'+
 									'<button class="remove-room-button">'+
@@ -83,6 +100,9 @@
 			'change:participantType': function() {
 				this.render();
 			},
+			'change:isFavorite': function() {
+				this.render();
+			},
 			'change:unreadMessages': function() {
 				this.render();
 			},
@@ -107,6 +127,7 @@
 			var isRemovable = this.model.get('type') !== 1;
 			return {
 				isRemovable: isRemovable,
+				canFavorite: this.model.get('participantType') !== 5,
 				isDeletable: !isRemovable || ((this.model.get('participantType') === 1 || this.model.get('participantType') === 2) &&
 					(Object.keys(this.model.get('participants')).length > 1 || this.model.get('numGuests') > 0)),
 				numUnreadMessages: this.model.get('unreadMessages') > 99 ? '99+' : this.model.get('unreadMessages')
@@ -133,6 +154,8 @@
 		},
 		events: {
 			'click .app-navigation-entry-utils-menu-button button': 'toggleMenu',
+			'click @ui.menu .favorite-room-button': 'addRoomToFavorites',
+			'click @ui.menu .unfavorite-room-button': 'removeRoomFromFavorites',
 			'click @ui.menu .remove-room-button': 'removeRoom',
 			'click @ui.menu .delete-room-button': 'deleteRoom',
 			'click @ui.room': 'joinRoom'
@@ -197,6 +220,30 @@
 
 			$.ajax({
 				url: OC.linkToOCS('apps/spreed/api/v1/room', 2) + this.model.get('token'),
+				type: 'DELETE'
+			});
+		},
+		addRoomToFavorites: function() {
+			if (this.model.get('participantType') === 5) {
+				return;
+			}
+
+			this.model.set('isFavorite', 1);
+
+			$.ajax({
+				url: OC.linkToOCS('apps/spreed/api/v1/room', 2) + this.model.get('token') + '/favorite',
+				type: 'POST'
+			});
+		},
+		removeRoomFromFavorites: function() {
+			if (this.model.get('participantType') === 5) {
+				return;
+			}
+
+			this.model.set('isFavorite', 0);
+
+			$.ajax({
+				url: OC.linkToOCS('apps/spreed/api/v1/room', 2) + this.model.get('token') + '/favorite',
 				type: 'DELETE'
 			});
 		},
